backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
jellyfish
rapidfuzz
sentence-transformers
faiss-cpu
torch
